requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-jose[cryptography]
apscheduler
twilio
email-validator
passlib==1.7.4
bcrypt==4.0.1
resend
twilio